Provide sphinx-*-{3.5,3} symlinks for each script
... following https://fedoraproject.org/wiki/Packaging:Python#Naming.
No need to put the full path in symlinks. It is actually better
to include only the last part of the path, because that the path
is resolved the same within and from outside of a chroot.

Rearange %files a bit to work when python3 is disabled.

Fix one more requirement on the python3 subpackage.